Adapt pick_next_task
if the prev belongs to ATLAS, call put_prev_task, because the following
evaluation requires it. If we decide we have an atlas task to run, we're good.
If we decide that we don't have a task and will return NULL, we need to revert
the action of calling put_prev_task, because the next scheduler having
something to run, will call put_prev_task again.
If we have a task to run, but put_prev_task was not an ATLAS task, then we have
to call put_prev_task. Maybe we can always call put_prev_task and revert the
effects later, regardless of the sched_class of prev.