BUGFIX slacktime
If we are in slacktime and not advancing, decrement nr_running and increment it
when the slack timer is up.
If we are advancing in CFS, the task is dequeued from ATLAS and the rn_running
count is correct.
If the count is not connect, CFS will detect it and think that a higher prio
sched_class has runnable task and instruct the scheduler to re-test all
sched_classes. Because we have not a runnable task, we will have an endless
loop.
The slack-timer will never fire, because it does not get run. The ATLAS slack
timer only tests against atlas_rq timer target, not the timeout value. Thus,
the test should probably be extended to include the timer value, and then ATLAS
will eventually have a runnable task.