Bugfix: Incomplete initialisation of sched_atlas_entity
In particular, se->job was not re-set, so that a child-thread can get a copy of
the atlas_job pointer from it's parent. This pointer, or an atlas_job allocated
at a later time at the same address might become invalid, when the child-thread
exits and deallocates "its" se->job. Subsequently, the atlas_job object of the
parent is corrupted. Detailed chain of events:
. parent has an atlas job.
. child thread spawns, se->job pointer is copied.
( the parent se->job might be deallocated and later an atlas job might be
reallocated to the same address, or the atlas_job is still valid)
. The client thread exits, deallocating the its se->job, leaving the parent with
  a dangling pointer
. Parent accesses se->job