Factor out sched_atlas_entity->real_job
sae->real_job was used to handle jobs which overran their deadline so far, that
they interfered with later jobs of the same task. In those cases, the task is
scheduled under ATLAS policy again for the duration of the later job, because
the time is accounted for in ATLAS anyway.

It was not really clear how sae->job and sae->real_job was used; neither why it
was necessary. The functionality is remains the same. In the new design,
sae->job points to the job on whose reservation the task is currently running
on. sae->job is strictly a reference and does not take ownership. The head of
sae->jobs is the job that is currently being processed by the task.

sae->jobs is the owner of an atlas_job, once a job is completed, it is removed
from the run queue (atlas_rq->jobs, rb-tree in deadline-order) and deleted from
the sae->jobs list. If a job misses its deadline it is removed from the run
queue but remains in the sae->jobs list, until atlas::next() signals the
completion of that job and it is removed.