Re-add atlas_do_pending_work()-pendant
An explicit function, called every schedule() call to move slack tasks back to
ATLAS. This requires touching the schedule() loop, but it avoids some
questionable stuff.
The edge-case is, when ATLAS has only one task, which overruns a job deadline
and is demoted to Recover or CFS. However, this task has future jobs
(reservations) which should be scheduled in ATLAS. Hence, pick_next_task takes
an early-out because nr_runnable is 0, but pick_next_task has to be called to
process atlas_set_scheduler, which would increase nr_runnable.
A hack was to temporarily increase nr_runnable in timer_rq_func with a dummy
and decrease in pick_next_task after atlas_set_scheduler had been called.