Reset scheduler to CFS before blocking in atlas_next()
Wakeups caused, for example, by SIGKILL (process exit), cause a resched.
Because when in Recover or ATLAS, the task does not have a job in RQ,
pick_next_task returns NULL, despite having an accounted ready task in
rq->nr_running. This leads CFS to try to re-schedule a higher-priority
sched_class and subsequently to an infinite scheduling loop.