Fix misuse of sched_atlas_entity.job
job is intended to point to the job, to which the current runtime is accounted
to. Previously, it was also used to distiguish the first call to next() from
subsequent ones. The first call is special, because no job may be deleted
(Because non has finished). This is fixed now, so job can be set to NULL when
put_prev_task is called.