Introduce jobs count for ATLAS RQ
Incoming patches the job count is going to replace nr_runnable and nr_running.
The redesign reflects the fact, that exactly one task is runnable at any given
time under ATLAS. This is a result of ATLAS' reserveration. As a consequence,
in_slack, etc is going to go away; simplifying nr_running management.