Bugfix: reset ATLAS_INIT only when returning w/ job
If the first call to next() blocked and blocking was interrupted by a signal,
the ATLAS_INIT flag would have been erroneously reset. The next call to next()
would cause destroy_first_job() to be called, which would either BUG() on not
having  a job or delete a job erroneously. Therefore, the ATLAS_INIT flag can
only be reset after next() returned with a job.