Change the meaning of atlas_job::[s]exectime/Introduce rexectime
Exectime is the amount of reserved CPU time and it is never changed, except by
atlas::update(). Sexectime is the amount of CPU time that the ATLAS scheduler
was able to reserve for the job (sexectime might be less than exectime in
overload situations, otherwise they are equal). rexectime is the amount of CPU
time, the Job already received, strictly increasing. As long as rexectime <
sexectime, the task is runnable. On the deadline, rexectime == sexectime is
true.