Bugfix: initalize last_cpu/mask after adding job
In schedule_job(), switched_from_atlas() is used to initialized last_cpu and
last_mask. However, it was called before the job was added to the jobs list. As
a result, if the task is still in CFS, and the first jobs is submitted, the
task would still be able to migrate in CFS.
Therefore,
1) switched_from_atlas() is only called if the task is currently in CFS
2) it is called after the job is added, such that migration in CFS is disabled
3) it is only necessary to call switched_from_atlas for the *first submission
   to an empty* list