Bugfix: Race condition between exit_atlas and submit
There was a window between exit_atlas() and detachment of the PID of the
exiting process, which allowed another thread to submit a job to the
now-exiting process.

This would result in the job not being properly cleaned up and ultimately
having a struct task_struct * pointing to an invalid task_struct.

Add an ATLAS_EXIT flag to atlas_se->flags when exit_atlas() is called, that is
checked on job submission. If that flag is set, the job is refused, pretending
the thread/process does not exist anymore.