Implement task_tick_atlas
just call update_curr_atlas().
Rescheduling is caused by timer and tasks being woken up (which causes
check_preempt_curr() to be called), so ticks are not necessary for ATLAS.
Updating task stats is done in pick_next_task()/put_task() anyway, so it is not
strictly necessary in task_tick().