Fix destroy_first_job with overload pulling
When overload pulling is enabled, next_job in destroy_first_job might already
have been pulled, so the BUG() macro is wrong.
Additionally, try to pull the next not-yet-pulled job, but do so conditionally
if the current run queue will not be overloaded by doing that.
Migrating back has to be done also conditionally. If there is no next job does
not mean that all jobs on the current run queue are done.