Fix migration in CFS
Fix migration in CFS and restricting it when ATLAS jobs are present.
Specifically, the edge cases of submitting an ATLAS job to a CFS task, removing
the last ATLAS job from a CFS task, and dropping into CFS with no job had to be
fixed. set_cpus_allowed() had to be modified to propagate sched_setaffinity()
calls to ATLAS in case ATLAS restricted the affinity of CFS tasks.